galery/requirements.txt
flask
flask_wtf
Pillow
boto3
python-dotenv
mysql-connector-python